Added values to time inputs to avoid null inputs
Added new select input in order to view multiple songs, created an event associated with selecting a song.

Modified original select to be associated with a specific model.

Created new method to search for songs based on original select input with the text input.

Error handling is not operating as expected
At an impass for seeding the database. Error in typing seems to be the initial cause.
Issues with deploying Heroku, maybe from the same issues that are rendering the server unusable.